⭐ New Features 1. menu 类根据前端代码修改成三级目录，以更好地实现权限控制，并增加 status 字段，permission 修改成 authCode。MenuTypeEnum 对应修改

+++ src/main/java/com/simonvonxcvii/turing/entity/Permission.kt
@@ -48,8 +48,8 @@ data class Permission(
     /**
      * 排序编号
      */
-    @Column(unique = true, nullable = false, columnDefinition = "SMALLINT", comment = "排序编号")
-    var sort: Short = 0,
+    @Column(unique = true, nullable = false, columnDefinition = "INTEGER", comment = "排序编号")
+    var sort: Int = 0,
 ) : AbstractAuditable() {
     companion object {
         /**
