+++ src/main/java/com/simonvonxcvii/turing/model/dto/PermissionDTO.java
@@ -36,7 +36,7 @@ public class PermissionDTO extends PageQuery {
      * 排序编号
      */
     @NotNull(message = "排序编号不能为空")
-    private Short sort;
+    private Integer sort;
     /**
      * 子级权限集合
      */
